patch: improve enforcement performance on legacy browsers
