+++ jsconfig.json
@@ -16,6 +16,7 @@
       "none": ["./packages/anyone/src/none.js"],
       "one": ["./packages/anyone/src/one.js"],
       "runAnyoneMethods": ["./packages/anyone/src/runner/runAnyoneMethods.js"],
+      "bindLazyRule": ["./packages/n4s/src/enforce/bindLazyRule.js"],
       "anyOf": ["./packages/n4s/src/enforce/compounds/anyOf.js"],
       "compounds": ["./packages/n4s/src/enforce/compounds/compounds.js"],
       "isArrayOf": ["./packages/n4s/src/enforce/compounds/isArrayOf.js"],
@@ -24,6 +25,8 @@
       "enforce": ["./packages/n4s/src/enforce/enforce.js"],
       "enforceKeywords": ["./packages/n4s/src/enforce/enforceKeywords.js"],
       "enforceRunner": ["./packages/n4s/src/enforce/enforceRunner.js"],
+      "genRuleProxy": ["./packages/n4s/src/enforce/genRuleProxy.js"],
+      "runtimeRules": ["./packages/n4s/src/enforce/runtimeRules.js"],
       "enforceExtended": ["./packages/n4s/src/extended/enforceExtended.js"],
       "rulesExtended": ["./packages/n4s/src/extended/rulesExtended.js"],
       "bindNot": ["./packages/n4s/src/lib/bindNot.js"],
