+++ packages/n4s/src/enforce/runtimeRules.js
@@ -1,0 +1,4 @@
+import compounds from 'compounds';
+import rules from 'rules';
+
+export default Object.assign(rules(), compounds);
